perf: 优化变量&server任务分步

+++ packages/core/build/remove.js
@@ -1,5 +1,7 @@
 const { rimraf } = require('rimraf')
 
-module.exports = function removeTask(dir, done) {
+module.exports = function removeTask(options = {}, done) {
+  const dir = typeof options === 'string' ? options : options.input
+  if (!dir) return
   rimraf(dir).then(() => done())
 }
